Revert "Separation of lessons per page."
This reverts commit a4d7d63a966c823f903eac4b89ec5dadd9f4dfb8.
