--- lesson-one/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="favicon.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>ThreeJS learning - Lesson one</title>
-  </head>
-  <body>
-    <a href="/">Back to the index</a>
-    <div class="lesson__result" id="lesson-result"></div>
-    <script type="module" src="/src/lesson-one/main.ts"></script>
-  </body>
-</html>
